+++ src/main/java/org/trebol/api/controllers/DataProductCategoriesController.java
@@ -3,15 +3,11 @@ package org.trebol.api.controllers;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import org.trebol.api.GenericDataController;
@@ -19,7 +15,6 @@ import org.trebol.api.DataPage;
 import org.trebol.api.pojo.ProductCategoryPojo;
 import org.trebol.config.CustomProperties;
 import org.trebol.jpa.entities.ProductCategory;
-import org.trebol.exceptions.EntityAlreadyExistsException;
 import org.trebol.jpa.GenericJpaCrudService;
 
 /**
@@ -49,18 +44,5 @@ public class DataProductCategoriesController
   @PreAuthorize("hasAuthority('product_categories:read')")
   public ProductCategoryPojo readOne(@PathVariable String code) {
     throw new UnsupportedOperationException("Method not implemented");
-  }
-
-  @Override
-  @ResponseStatus(HttpStatus.BAD_REQUEST)
-  @ExceptionHandler(MethodArgumentNotValidException.class)
-  public Map<String, String> handleException(MethodArgumentNotValidException ex) {
-    return super.handleException(ex);
-  }
-
-  @ResponseStatus(HttpStatus.BAD_REQUEST)
-  @ExceptionHandler(EntityAlreadyExistsException.class)
-  public String handleException(EntityAlreadyExistsException ex) {
-    return ex.getMessage();
   }
 }
