+++ src/main/java/org/trebol/api/controllers/PublicCheckoutController.java
@@ -5,8 +5,6 @@ import java.util.Map;
 
 import javax.validation.Valid;
 
-import io.jsonwebtoken.Claims;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -18,20 +16,18 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import org.trebol.api.pojo.PaymentRedirectionDetailsPojo;
 
+import org.trebol.api.pojo.PaymentRedirectionDetailsPojo;
 import org.trebol.api.pojo.SellPojo;
 import org.trebol.exceptions.BadInputException;
 import org.trebol.integration.exceptions.PaymentServiceException;
+import org.trebol.api.ICheckoutService;
 
 import javassist.NotFoundException;
 
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
-import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.SEE_OTHER;
 
-import org.trebol.api.ICheckoutService;
 
 /**
  *
@@ -99,25 +95,9 @@ public class PublicCheckoutController {
     return service.getResultingTransaction(token);
   }
 
-  @ResponseStatus(BAD_REQUEST)
-  @ExceptionHandler(BadInputException.class)
-  public String handleException(BadInputException ex) {
-    return ex.getMessage();
-  }
-
   @ResponseStatus(INTERNAL_SERVER_ERROR)
   @ExceptionHandler(PaymentServiceException.class)
   public String handleException(PaymentServiceException ex) {
-    return ex.getMessage();
-  }
-
-  @ResponseStatus(NOT_FOUND)
-  @ExceptionHandler(NotFoundException.class)
-  public void handleException(NotFoundException ex) { }
-
-  @ResponseStatus(INTERNAL_SERVER_ERROR)
-  @ExceptionHandler(RuntimeException.class)
-  public String handleException(RuntimeException ex) {
     return ex.getMessage();
   }
 }
