+++ src/main/java/org/trebol/api/controllers/PublicProductsController.java
@@ -3,16 +3,13 @@ package org.trebol.api.controllers;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import org.trebol.api.DataPage;
 
+import org.trebol.api.DataPage;
 import org.trebol.api.pojo.ProductPojo;
 import org.trebol.config.CustomProperties;
 import org.trebol.jpa.GenericJpaCrudService;
@@ -47,8 +44,4 @@ public class PublicProductsController {
   public ProductPojo readOne(@PathVariable Long id) throws NotFoundException {
     return productsCrudService.readOne(id);
   }
-
-  @ResponseStatus(HttpStatus.NOT_FOUND)
-  @ExceptionHandler(NotFoundException.class)
-  public void handleException(NotFoundException ex) { }
 }
